+++ packages/tools/utils/CHANGELOG.md
@@ -9,9 +9,14 @@
 
 - renamed `makeIsContained` to `makeOccursIn`
 
+### Object -> (Object -> Object)
+
+- `transformValues`: assume identity for missing keys
+
 ### Object -> Object
 
 - add `swapKeyValue`
+
 
 ## `@svizzle/utils` v0.5.0